+++ backend/routes/auth.js
@@ -3,7 +3,7 @@ const bcrypt = require('bcryptjs');
 const jwt = require('jsonwebtoken');
 const crypto = require('crypto');
 const { body, validationResult } = require('express-validator');
-const User = require('../models/User');
+const User = require('../models/user/user.model');
 const { sendEmail } = require('../utils/emailService');
 const { asyncHandler } = require('../middleware/errorHandler');
 const { authenticateToken } = require('../middleware/auth');
