+++ backend/server.js
@@ -1,23 +1,30 @@
-const express = require('express');
-const mongoose = require('mongoose');
-const cors = require('cors');
-const helmet = require('helmet');
-const rateLimit = require('express-rate-limit');
+// ✅ Load environment variables
 const dotenv = require('dotenv');
-const http = require('http');
-const socketIo = require('socket.io');
-
-// Load environment variables
 dotenv.config();
 
-// Import routes from routes/index.js
-const routes = require('./routes');
+// ✅ Core modules
+const http = require('http');
+const mongoose = require('mongoose');
+const socketIo = require('socket.io');
 
-// Import middleware
-const { errorHandler } = require('./middleware/errorHandler');
-const { authenticateToken } = require('./middleware/auth');
+// ✅ Import express app
+const app = require('./app');
 
-const app = express();
+// ✅ Config variables
+const PORT = process.env.PORT || 5000;
+const MONGO_URI = process.env.MONGO_URI || 'mongodb://localhost:27017/smart-healthcare';
+
+// ✅ Connect MongoDB
+mongoose.connect(MONGO_URI, {
+  useNewUrlParser: true,
+  useUnifiedTopology: true
+}).then(() => {
+  console.log("✅ MongoDB connected");
+}).catch(err => {
+  console.error("❌ MongoDB connection failed:", err);
+});
+
+// ✅ Create HTTP server & bind Socket.io
 const server = http.createServer(app);
 const io = socketIo(server, {
   cors: {
@@ -26,77 +33,15 @@ const io = socketIo(server, {
   }
 });
 
-// Security middleware
-app.use(helmet());
-app.use(cors({
-  origin: process.env.FRONTEND_URL || "http://localhost:3000",
-  credentials: true
-}));
-
-// Rate limiting
-const limiter = rateLimit({
-  windowMs: 15 * 60 * 1000,
-  max: 100
-});
-
-const authLimiter = rateLimit({
-  windowMs: 15 * 60 * 1000,
-  max: 50,
-  message: 'Too many login attempts, please try again later.'
-});
-
-// Apply rate limiters
-app.use('/api/', limiter);
-app.use('/api/auth/', authLimiter);
-
-// Body parsing middleware
-app.use(express.json({ limit: '10mb' }));
-app.use(express.urlencoded({ extended: true, limit: '10mb' }));
-
-// Serve static files
-app.use('/uploads', express.static('uploads'));
-
-// MongoDB connection
-mongoose.connect(process.env.MONGO_URI || 'mongodb://localhost:27017/smart-healthcare', {
-  useNewUrlParser: true,
-  useUnifiedTopology: true
-});
-
-// Socket.io connection
+// ✅ Handle socket connection logic
 require('./socket/chatSocket')(io);
 
-// Conditional token middleware: applies token check to all except /auth
-function authenticateTokenUnlessAuth(req, res, next) {
-  if (req.path.startsWith('/auth')) return next();
-  authenticateToken(req, res, next);
-}
-
-// Mount all routes
-app.use('/api', authenticateTokenUnlessAuth, routes);
-
-// Health check endpoint
-app.get('/api/health', (req, res) => {
-  res.json({
-    status: 'OK',
-    message: 'Smart Healthcare Assistant API is running',
-    timestamp: new Date().toISOString()
-  });
-});
-
-// Error handler
-app.use(errorHandler);
-
-// 404 handler
-app.use('*', (req, res) => {
-  res.status(404).json({ message: 'Route not found' });
-});
-
-const PORT = process.env.PORT || 5000;
-
+// ✅ Start server
 server.listen(PORT, () => {
   console.log(`🚀 Server running on port ${PORT}`);
   console.log(`📱 Frontend URL: ${process.env.FRONTEND_URL || 'http://localhost:3000'}`);
   console.log(`🔗 API URL: http://localhost:${PORT}/api`);
 });
 
+// ✅ Export (for testing or other purposes)
 module.exports = { app, server, io };
